[added] system message for weapon decay [fixed] getConditionDamage now returns int (the float is still used to store damage)
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@4383 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/objects/tangible/TangibleObject.h
@@ -264,7 +264,7 @@ public:
 
 	void setMaxCondition(int maxCond, bool notifyClient = true);
 
-	float getConditionDamage();
+	int getConditionDamage();
 
 	int getVolume();
 
@@ -487,7 +487,7 @@ public:
 
 	void setMaxCondition(int maxCond, bool notifyClient = true);
 
-	float getConditionDamage();
+	int getConditionDamage();
 
 	int getVolume();
 
@@ -674,7 +674,7 @@ public:
 
 	void setMaxCondition(int maxCond, bool notifyClient);
 
-	float getConditionDamage();
+	int getConditionDamage();
 
 	int getVolume();
